chore: Move info parsing related method from OC_App to InfoParser
Also fix the tests
